Move test client to clients.py
